+++ src/entry.esm.js
@@ -1,16 +1,15 @@
-
 // Import vue component
 import component from '@/litepie-datepicker.vue';
 
 // Default export is installable instance of component.
 // IIFE injects install function into component, allowing component
 // to be registered via Vue.use() as well as Vue.component(),
-export default /*#__PURE__*/(() => {
+export default /*#__PURE__*/ (() => {
   // Get component instance
   const installable = component;
 
   // Attach install function executed by Vue.use()
-  installable.install = (app) => {
+  installable.install = (app, options) => {
     app.component('LitepieDatepicker', installable);
   };
   return installable;