+++ CHANGELOG.md
@@ -47,6 +47,13 @@ All notable changes to this project will be documented in this file.
 - Change local model `value` to `pickerValue`.
 - Fix issues placement arrow, `left` to `right`.
 
+## [1.0.9]
+
+- Props `footer` removed.
+- Fix localization
+- Fix bug `auto-apply` if shortcuts clicked.
+- Add new props `options`, [Example](https://litepie.com/#options).
+
 [Released]: https://github.com/kenhyuwa/litepie-datepicker/
 [1.0.0]: https://github.com/kenhyuwa/litepie-datepicker/releases/tag/v1.0.0
 [1.0.1]: https://github.com/kenhyuwa/litepie-datepicker/releases/tag/v1.0.1
@@ -57,3 +64,4 @@ All notable changes to this project will be documented in this file.
 [1.0.6]: https://github.com/kenhyuwa/litepie-datepicker/releases/tag/v1.0.6
 [1.0.7]: https://github.com/kenhyuwa/litepie-datepicker/releases/tag/v1.0.7
 [1.0.8]: https://github.com/kenhyuwa/litepie-datepicker/releases/tag/v1.0.8
+[1.0.9]: https://github.com/kenhyuwa/litepie-datepicker/releases/tag/v1.0.9